requirements.txt
Flask
flask_sqlalchemy
flask_migrate
flask_login
flask_wtf
email-validator
gunicorn
requests

# JWT Authentication
flask-jwt-extended

# API Serialization & Validation
marshmallow
marshmallow-sqlalchemy

# CORS Support for Mobile Apps
flask-cors

# Testing
pytest
